HABA-735 Data collection file handling
* Only create data collection files if needed
* Handle no data-collection files on first boot

+++ recipes-flir/skylab-files-mgmt/files/skylab-files-mgmt.sh
@@ -20,7 +20,7 @@
 # /tmp/skylab/CameraFiles.zip         -> /FLIR/images/skylab/CameraFiles.zip
 # /tmp/skylab/LicenceFiles.zip        -> /FLIR/images/skylab/LicenceFiles.zip
 # /tmp/skylab/DataCollectionFiles.zip -> /FLIR/images/skylab/DataCollectionFiles.zip
-# 
+#
 # Datacollection files are zipped on every boot. The original files
 # are deleted after the zip is created.
 #
@@ -43,14 +43,26 @@ EXCLUDED_FILES=statistics/active.log
 CREDENTIALS_FILE=credentials.json
 DATACOLLECTION_UUID_FILENAME_GLOB="????????-????-????-????-????????????"
 
-# DESCRIPTION: Extract field "storage_path" in section set by
-#              argument. It looks for json content on stdin.
-# USAGE:       get_storage_path <SECTION> < JSON_FILE
-# RETURNS:     basename of the value of storage_part
+# DESCRIPTION: Reads in json data on stdin and returns the value of
+#              the field SECTION/storage_path where SECTION is an
+#              argument.
+# USAGE:       get_storage_path <SECTION> < <JSON_FILE>
+# RETURNS:     basename of the value of storage_path
+#
+# EXAMPLE:     file.json contains:
+#              {
+#                  "minidump": {
+#                      "storage_path": "/some/path"
+#                  }
+#              }
+#
+#              Running the following will return "/some/path":
+#              get_storage_path "minidump:" < file.json
 function get_storage_path {
     IN_SECTION=""
     SECTION=$1
-    
+
+    # read from stdin
     while read CONF_LINE
     do
 	if [[ "$CONF_LINE" =~ $SECTION ]]
@@ -60,24 +72,44 @@ function get_storage_path {
 	fi
 
 	if [[ "$IN_SECTION" == "yes" && "$CONF_LINE" =~ storage_path ]]
-	then 
+	then
 	    basename "$(echo "$CONF_LINE" | cut -d: -f2 | tr -d ',"')"
 	    return
 	fi
     done
 }
 
+#
+# Create a zip file containing calibration files. There is a file on
+# the device that specifies all calibration files to include in the
+# zip file. The file is structured into 3 columns:
+#
+# filename | dir on filesystem | destination dir in zip file
+#
+# This function creates symbolic links to each file on the device in a
+# temporary directory. This temporary directory is then zipped in
+# persistent storage and exposed to skylab.
+#
+# example:
+# calib.rsc /FLIR/system CameraFiles/system
+#
+# CameraFiles/system/FLIR/system/calib.rsc -> /FLIR/system/calib.rsc
+#
+# CameraFiles/system is then zipped into CameraFiles.zip. Finally a
+# link in /srv/skylab is created pointing to the zip file.
+#
 function camera_files() {
     if [ ! -f "$LIST_OF_CALIB_FILES" ]
     then
 	echo "File $LIST_OF_CALIB_FILES not found"
 	return 1
     fi
-    
+
     cd /tmp || return 1
 
-    # ZIP_DIR is a directory within CAMERAFILES_TMPDIR.
-    mkdir "$CAMERAFILES_TMPDIR"
+    # The following code assumes that ZIP_DIR are directories within
+    # CAMERAFILES_TMPDIR/system.
+
     while read FILENAME FS_DIR ZIP_DIR
     do
 	if [ -e "$FS_DIR/$FILENAME" ]
@@ -85,32 +117,51 @@ function camera_files() {
 	    [ -d "$ZIP_DIR" ] || mkdir -p "$ZIP_DIR"
 	    ln -s "$FS_DIR/$FILENAME" "$ZIP_DIR"
 	else
-	    echo "$FS_DIR/$FILENAME missing"
+	    echo "Calib file $FS_DIR/$FILENAME not found"
 	fi
     done < "$LIST_OF_CALIB_FILES"
 
-    [ -e "$CAMERAFILES_ZIP" ] && rm "$CAMERAFILES_ZIP"
-    
     dir=$(dirname "$CAMERAFILES_ZIP")
     if [ -z "$dir" ]
     then
-	echo "failed to get directory of camerafiles"
+	echo "failed to get directory of camerafiles zip file"
 	return 1
     fi
-
     [ -d "$dir" ] || mkdir -p "$dir"
-    zip -r "$CAMERAFILES_ZIP" "${CAMERAFILES_TMPDIR}/system"
-    ln -sf "$CAMERAFILES_ZIP" "${VOLATILE_STORAGE}/$CAMERAFILES_FILENAME"
+
+    rm -f "$CAMERAFILES_ZIP"
+    zip -r "$CAMERAFILES_ZIP" "$CAMERAFILES_TMPDIR/system"
+
+    # Make the zip file available to skylab
+    ln -sf "$CAMERAFILES_ZIP" "$VOLATILE_STORAGE/$CAMERAFILES_FILENAME"
+
+    # Clean up
     rm -rf "$CAMERAFILES_TMPDIR"
 }
 
 function licence_files() {
-    if [ ! -e "$LICENCEFILES_ZIP" ]
+    cd /usr/share/common-licenses || return 1
+
+    rm -f "$LICENCEFILES_ZIP"
+    zip "$LICENCEFILES_ZIP" license.manifest licenses.html
+
+    # Make the zip file available to skylab
+    ln -sf "$LICENCEFILES_ZIP" "$VOLATILE_STORAGE/$LICENCEFILES_FILENAME"
+}
+
+function zip_data_collection_files() {
+    if [ -z "$1" ]
     then
-	cd /usr/share/common-licenses || { echo directory /usr/share/common-licenses missing -- exit; return 1; }
-	zip "$LICENCEFILES_ZIP" license.manifest licenses.html
+	echo "$1 do not exist"
+    else
+	while read -d '' f
+	do
+	    zip -r "$DATACOLLECTIONFILES_ZIP" "$f" -x $EXCLUDED_FILES
+	done < <(find "$1" \
+		      -type f \
+		      -regex '.*/[a-zA-Z0-9]\{8\}-[a-zA-Z0-9]\{4\}-[a-zA-Z0-9]\{4\}-[a-zA-Z0-9]\{4\}-[a-zA-Z0-9]\{12\}' \
+		      -print0)
     fi
-    ln -sf "$LICENCEFILES_ZIP" "${VOLATILE_STORAGE}/$LICENCEFILES_FILENAME"
 }
 
 function datacollection_files() {
@@ -122,32 +173,28 @@ function datacollection_files() {
 	return 1
     fi
 
+    cd /FLIR/system/data-collection/ || return 1
+
     MINIDUMP_STORAGE_PATH=$(eval echo "$(get_storage_path '"minidump":' < $DATA_COLLECTION_CONF)")
-    if [ -z "$MINIDUMP_STORAGE_PATH" ]
-    then
-	echo "Failed to find storage path to minidump"
-	return 1
-    fi
+    zip_data_collection_files "$MINIDUMP_STORAGE_PATH"
 
     STATISTICS_STORAGE_PATH=$(eval echo "$(get_storage_path '"statistics":' < $DATA_COLLECTION_CONF)")
-    if [ -z "$STATISTICS_STORAGE_PATH" ]
-    then
-	echo "Failed to find storage path to statistics"
-	return 1
-    fi
+    zip_data_collection_files "$STATISTICS_STORAGE_PATH"
 
-    STUFF_TO_STORE="$MINIDUMP_STORAGE_PATH $STATISTICS_STORAGE_PATH $CREDENTIALS_FILE"
-    cd /FLIR/system/data-collection/ || { echo directory /FLIR/system/data-collection/ missing -- exit; return 1; }
-    zip -r "$DATACOLLECTIONFILES_ZIP" $STUFF_TO_STORE -x $EXCLUDED_FILES
+    # return if no data collection files were zipped
+    [ -f "$DATACOLLECTIONFILES_ZIP" ] || return 1
+
+    zip -r "$DATACOLLECTIONFILES_ZIP"  "$CREDENTIALS_FILE"
 
     # Make the datacollection zip file available to skylab
     ln -sf "$DATACOLLECTIONFILES_ZIP" "${VOLATILE_STORAGE}/$DATACOLLECTIONFILES_FILENAME"
 
     # Cleanup old dump files
-    rm -f "$MINIDUMP_STORAGE_PATH/"$DATACOLLECTION_UUID_FILENAME_GLOB "$STATISTICS_STORAGE_PATH/"$DATACOLLECTION_UUID_FILENAME_GLOB
+    rm -f "$MINIDUMP_STORAGE_PATH/"$DATACOLLECTION_UUID_FILENAME_GLOB \
+       "$STATISTICS_STORAGE_PATH/"$DATACOLLECTION_UUID_FILENAME_GLOB
 }
 
-# .fuf files now end up in /tmp - clean out any remaining .fuf files
+# clean out any old .fuf files
 rm -f /FLIR/images/skylab/*.fuf
 
 # Create a directory in volatile storage and make it accessible to
@@ -155,14 +202,8 @@ rm -f /FLIR/images/skylab/*.fuf
 [ -d "$VOLATILE_STORAGE" ] || mkdir "$VOLATILE_STORAGE"
 ln -sf "$VOLATILE_STORAGE" /srv
 
-res=0
 camera_files
-res=$(( res + $? ))
-
 licence_files
-res=$(( res + $? ))
-
 datacollection_files
-res=$(( res + $? ))
 
-exit $res
+exit 0
